+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/dataset/ITable.java
@@ -219,7 +219,7 @@ public interface ITable<T> extends IMatrix<T> {
     }
 
     @Override
-    default int[] getShape() {
+    default int[] getSize() {
         return new int[]{getColumnCount(), getRowCount()};
     }